transfered js to js file, svg folder, service
