+++ src/app/Http/Controllers/UserController.php
@@ -4,18 +4,23 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use App\Models\User;
+use App\Services\UserService;
 use Illuminate\View\View;
 
 class UserController extends Controller
 {
+    protected $userService;
+    public function __construct(UserService $userService){
+        $this->userService = $userService;
+    }
+
     /**
      * Show other user's profile
      */
 
     public function show(int $id): View
     {
-        $user = User::with('profile')->findOrFail($id);
-        $posts = Post::where('user_id', $user->id)->get();
-        return view('pages.profile.profile', compact('user', 'posts'));
+        $data = $this->userService->getUserProfile($id);
+        return view('pages.profile.profile', $data);
     }
 }
