+++ src/resources/js/app.js
@@ -1,3 +1,4 @@
 import './bootstrap';
 import './buttons';
 import './hidden-menu';
+import './img-prev';
