+++ src/resources/views/pages/dashboard/post-content.blade.php
@@ -17,9 +17,7 @@
                     <a class="hover:text-white" href="{{route('view.post', ['post' => $post->id])}}">Edit</a>
               </div>
               <span style="cursor:pointer;" onclick="openNav({{$post->id}})">
-                  <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                      <path stroke-linecap="round" stroke-linejoin="round" d="M12 6.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5ZM12 12.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5ZM12 18.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5Z" />
-                  </svg>
+              @include('svg.kebab')
               </span>
           @endif
       </div>
@@ -46,9 +44,7 @@
 
       <span href="" class="d-dashboard__like-comm cursor-pointer" onclick="openComment({{$post->id}})">
                 <span class="flex justify-center items-center gap-2">
-                    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="currentColor" class="w-7 h-7 fill-mygray">
-                        <path fill-rule="evenodd" d="M5.337 21.718a6.707 6.707 0 0 1-.533-.074.75.75 0 0 1-.44-1.223 3.73 3.73 0 0 0 .814-1.686c.023-.115-.022-.317-.254-.543C3.274 16.587 2.25 14.41 2.25 12c0-5.03 4.428-9 9.75-9s9.75 3.97 9.75 9c0 5.03-4.428 9-9.75 9-.833 0-1.643-.097-2.417-.279a6.721 6.721 0 0 1-4.246.997Z" clip-rule="evenodd" />
-                    </svg> 
+                  @include('svg.comment')
                   Comment
               </span>
             </span>
