+++ src/resources/views/svg/phone.blade.php
@@ -1,0 +1,7 @@
+<svg class="h-7 w-7 text-black-700 mr-2" viewBox="0 0 24 24" stroke-width="1" stroke="currentColor" fill="none"
+    stroke-linecap="round" stroke-linejoin="round">
+    <path stroke="none" d="M0 0h24v24H0z" />
+    <path d="M5 4h4l2 5l-2.5 1.5a11 11 0 0 0 5 5l1.5 -2.5l5 2v4a2 2 0 0 1 -2 2a16 16 0 0 1 -15 -15a2 2 0 0 1 2 -2" />
+    <path d="M15 7a2 2 0 0 1 2 2" />
+    <path d="M15 3a6 6 0 0 1 6 6" />
+</svg>
